Integrate the MALI module into the kernel DEB file.
Instead of installing the Mali modules directly into the SD card image install
it into the kernel build structure from where the DEB package is being built.
Thus the mali module is included in the kernel DEB, and being installed in one step, and can be re-installed, and cannot get lost, when you switch kernels.