makenewimage.sh: Allow build for lima and KMS.
Add parameter --no-mali.
This still builds and installs the out-of-tree Mali driver but it is
blacklisted.
Furthermore the Mali blob is not installed. Instead the Mesa EGL, GLES 2.0,
and DRM headers and libs are installed instead.
You still can switch back to the ARM Mali by remove the blacklisting of the
mali module, blacklist lima, remove the Mesa packages, and install the Mali blob. The .deb packages for the Mali blob are still available.

Write the Mesa development packages into /mesa-dev-packages.txt for later
installation on the target machine.

Automatially activate the Ethernet interface etho, and automatically activate
a USB-tethered mobile device (e.g. Android phone) as interface usb0.

Replace package avahi-discover with avahi-utils. avahi-discover is a graphical
frontend, and installs all X and GTK, and Gnome components which I do not want.