makenewimage.sh: Mount file systems with option "noatime"
Without this option any access to a file or directory causes a write access to
update the atime (access time) of the i-node.
Particularly on SD cards this may lead to wear, and unnecessary disk accesses.
atime is used by some mailers and some backup programs. It is mostly obsolete.
This is no concern for the foreseen purpose.