makenewimage.sh: Create symbolic link for device tree based on selection.
The default Cubieboard 2 device tree, and the OpenVario device tree files
are all copied to sdcard/boot.
By default the default Cubie device tree is linked to /boot/openvario.dtb.
Then the used can interactively select if she wants to select another
device tree for one of the supported LVDS panels.
For the 5" and 7" Chefree (Texim is distributer, not manufacturer)
device trees for the redesigned OpenVario DS2 are available too.